src/requirements.txt
openai
anthropic
google-genai
pandas
numpy
tqdm
python-dotenv
matplotlib
opencv-python
langchain-community
crewai